Big blueprint/dungeon room refactor. Make dungeon room registry be THE room registry, and blueprints only have base blueprint stuff. Make things like door locations, chest locations, etc. be a DungeonRoom thing, NOT a blueprint thing.
One side-effect of this is placeholder blocks like entry/exit spots are
still in the blueprints now (and I've regenerated all the rooms to do
this) and to blueprints they are just redstone repeaters/comparators.

Rooms treat them special and note their location / mark the entry when
capturing, but blueprints don't care.